chore: rearrange test structure to allow for better fixture usage
